# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIhidalgo.devs",
        "description": "Api escuela hidalgo",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "Equipo de soporte",
            "email": "msolegario.cisneros@gmail.com"
        },
        "license": {
            "name": "Licencia",
            "url": "https://example.com/license"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://por_definir/api/documentation"
        }
    ],
    "paths": {
...
        "/api/califications": {
            "post": {
                "tags": [
                    "califications"
                ],
                "summary": "Add a new calification",
                "description": "Create a new calification for a student in a specific activity.",
                "operationId": "addCalification",
                "requestBody": {
                    "description": "Calification data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "activity_id",
                                    "student_id",
                                    "calification"
                                ],
                                "properties": {
                                    "activity_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "student_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "calification": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 8.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Calification created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "created"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "data": {
                                            "properties": {
                                                "activity_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "student_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "calification": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 8.5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "bad request"
                                        },
                                        "error": {
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "internal server error"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error message"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}